backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic==2.5.2
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

python-multipart==0.0.6
google-generativeai==0.8.6
python-dotenv==1.0.0
sqlalchemy==2.0.23
httpx==0.25.1
